4.2

+++ PrivateFrameworks/Search.framework/SPPlistBytesDeserialization.h
@@ -4,7 +4,7 @@
 
 @interface SPPlistBytesDeserialization : NSObject {
     char *_buffer;
-    NSUInteger _length;
+    unsigned long _length;
 }
 
 - (void)dealloc;
